refactor: steps 폴더로 이동
